Rename files to model/view/etc.
