+++ corrscope/gui/view_stack.py
